requirements.txt
transformers==4.37.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
langchain==0.1.0
langchain-community==0.0.13
pypdf2==3.0.1
streamlit==1.32.0
python-dotenv==1.0.0
pydantic==2.4.2
fastapi==0.105.0
uvicorn==0.24.0
nltk==3.8.1
chardet==5.2.0
huggingface_hub==0.20.2
python-docx==1.1.0